Change core write(), append() ready for batch-method.
write(filename,...) now succeeds and returns true iff
filename does not exist but its dir does exist.
append(filename,...) now succeeds and returns true iff
filename exists and, of course, its dir exists too.
Idea is to create a batch method that accepts a sequence
of any method calls. Eg
[
  ['exist?','/cyber-dojo/kata/3e/f5/6s/'],
  ['make?', '/cyber-dojo/kata/3e/f5/6s/3'],
  ['write', '/cyber-dojo/kata/3e/f5/6s/3/event.json', '...'],
  ['append','/cyber-dojo/kata/3e/f5/6s/events.json', '...']
]
This batch method will stop as soon as an individual
method returns false.

+++ src/saver.rb
@@ -21,19 +21,27 @@ class Saver
 
   # - - - - - - - - - - - - - - - - - - - - - - - -
 
-  def append(filename, content)
-    make?(File.dirname(filename))
-    File.open(filename, 'a') { |fd| fd.write(content) }
-  end
-
   def write(filename, content)
-    make?(File.dirname(filename))
-    File.open(filename, 'w') { |fd| fd.write(content) }
+    if Dir.exist?(File.dirname(filename)) && !File.exist?(filename)
+      File.open(filename, 'w') { |fd| fd.write(content) }
+      true
+    else
+      false
+    end
   end
 
-  def reads(filenames) # read() BatchMethod
-    filenames.map{ |filename| read(filename) }
+  # - - - - - - - - - - - - - - - - - - - - - - - -
+
+  def append(filename, content)
+    if Dir.exist?(File.dirname(filename)) && File.exist?(filename)
+      File.open(filename, 'a') { |fd| fd.write(content) }
+      true
+    else
+      false
+    end
   end
+
+  # - - - - - - - - - - - - - - - - - - - - - - - -
 
   def read(filename)
     if File.file?(filename)
@@ -41,6 +49,12 @@ class Saver
     else
       nil
     end
+  end
+
+  # - - - - - - - - - - - - - - - - - - - - - - - -
+
+  def reads(filenames) # read() BatchMethod
+    filenames.map{ |filename| read(filename) }
   end
 
 end
